fix: Ensure deleteItem handles deleting root node
this.root is initially passed in as a parameter called 'root' to
deleteItemRecursive. This allows the original this.root object to be
mutated, but trying to set root = null or similar has no effect on the
actual this.root property, due to how Javascript works with variables:
https://medium.com/@anthonygood/pass-by-sharing-in-javascript-and-why-it-matters-829c96163650

In the deletion case where root has two children this works fine, as
only the .value property is changed (e.g. mutation of the original
object). However, deleting a root node with one or zero children
requires setting this.root to a new object, not just mutating the
original.

To fix this, set this.root to the output of deleteItemRecursive, which
returns the (modified if needed) root node at the end of its execution.