fix: Ensure insert correctly handles empty tree
this.root is initially passed in as a parameter called 'root' to
insertRecursive. This allows the original this.root object to be
mutated, but poses some issues when the root itself is null (an empty
tree). Trying to set root = new Node(value) or similar has no effect on
the actual this.root property due to how Javascript works
with variables:
https://medium.com/@anthonygood/pass-by-sharing-in-javascript-and-why-it-matters-829c96163650

This means when the root itself is null, the insertRecursive function
will attempt to read a property from the root, which causes an error.

To fix this, check beforehand if the root is null. If it is, set it to
a new node with the given value and start the tree, then return.